Splits MessageList into smaller components
