+++ client/src/Components/ConversationDetail.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Col, Panel, Grid, Image, Label, ListGroup, ListGroupItem, PageHeader, Row, Tab, Table, Tabs } from 'react-bootstrap';
 import Moment from 'react-moment';
 import PropTypes from 'prop-types';
-import MessageList from './MessageList';
+import MessageList from './MessageList/MessageListContainer';
 import HttpRequest from './HttpRequest';
 
 const helpers = require('../helpers');
