+++ client/src/Components/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Grid } from 'react-bootstrap';
-import MessageList from './MessageList';
+import MessageList from './MessageList/MessageListContainer';
 
 const Home = () => (
   <Grid>
